feat: Add Blackhole experiment, model selector UI, and optimizations
- Add Blackhole experiment for multi-path semantic exploration
- Add model selector dropdown in header for runtime model switching
- Add /api/models and /api/models/switch endpoints
- Optimize Prism by removing slow LLM refinement step
- Add detailed timing logs to all experiments (Prism, Blackhole, Wormhole)
- Strip <think> tokens from LLM output for Thinking model variants
- Experiment grid UI replacing vertical tabs
- Fix wormhole syntax error and model cache issue

+++ backend/latent_cartographer.py
@@ -187,12 +187,12 @@ class LatentCartographer:
             # Handle different shapes
             if len(act.shape) == 3:
                 # [batch, seq, hidden]
-                vec = act[0, token_idx, :].cpu().numpy()
+                vec = act[0, token_idx, :].float().cpu().numpy()
             elif len(act.shape) == 2:
                 # [seq, hidden]
-                vec = act[token_idx, :].cpu().numpy()
+                vec = act[token_idx, :].float().cpu().numpy()
             else:
-                vec = act.cpu().numpy().flatten()
+                vec = act.float().cpu().numpy().flatten()
                 
             # Project to 3D
             coords = self.project(vec)
